+++ mvd-analytics/analyzer/postprocess.go
@@ -101,6 +101,17 @@ func normalizeMatchRelativeTimes(res *Result, _ *CoreOutputs) {
 		for mi := range streams.Movers {
 			shiftAndClampMoverStream(&streams.Movers[mi], matchStartMs)
 		}
+		if pr := streams.Projectiles; pr != nil {
+			for i := range pr.Spawn {
+				pr.Spawn[i] -= matchStartMs
+				pr.End[i] -= matchStartMs
+			}
+		}
+		if bm := streams.Beams; bm != nil {
+			for i := range bm.T {
+				bm.T[i] -= matchStartMs
+			}
+		}
 	}
 
 	if res.Messages != nil {